Rename files using snake-case and adding ui folder for base components
