+++ src/app/dashboard/page.tsx
@@ -1,5 +1,3 @@
-import { Text } from "@/components/base/Text";
-
 export default function Dashboard() {
   return <></>;
 }
